new event page added
