--- src/New events/app/layout.jsx
@@ -1,14 +1,0 @@
-import "./globals.css";
-
-export const metadata = {
-    title: "Events",
-    description: "Premium Events Platform",
-};
-
-export default function RootLayout({ children }) {
-    return (
-        <html lang="en">
-            <body>{children}</body>
-        </html>
-    );
-}
