--- src/New events/app/page.jsx
@@ -1,13 +1,0 @@
-"use client";
-
-import Hero from "@/components/sections/hero";
-import Footer from "@/components/sections/footer";
-
-export default function EventsPage() {
-    return (
-        <main className="min-h-screen bg-[#0a0e17] overflow-x-hidden">
-            <Hero />
-            <Footer />
-        </main>
-    );
-}
